Added random card functionality
